Change the header of the pages
